refactor: extract Kilo model catalog mapping
Move pure catalog mapping into a dedicated module with direct coverage for gateway thinking variants. Update the mapping to support Pi's max thinking level; this also exposes the documented DeepSeek V4 max variant when the gateway metadata omits it.

Fixes #3
